direct music

+++ src/Audio/AudioPlayer.h
@@ -80,8 +80,8 @@ public:
 protected:
 	AudioDevice*		mDevice;
 	AudioStream*		m_audioStream;
-    float			    mVolume;		///< 音量 ( 0 ～ 100 )
-	float			    mPitch;			///< ピッチ ( 50 ～ 200 )
+    int				    mVolume;		///< 音量 ( 0 ～ 100 )
+	int				    mPitch;			///< ピッチ ( 50 ～ 200 )
 	uint32_t			mLoopBegin;		///< ループされる領域の最初のサンプル (Midi なら ミュージックタイム単位)
 	uint32_t			mLoopLength;	///< ループ領域の長さ (サンプル数単位)  (Midi なら ミュージックタイム単位)
 	bool			    mIsPlaying;
